mongo_to_sql/requirements.txt
confluent_kafka
pymongo
pydantic
pydantic-settings
mysql-connector-python
